+++ src/Makevars
@@ -1,7 +1,3 @@
 # -*- Makefile -*-
 
-# Link with GMP  (GNU Multiple Precision)
-# and  with MPFR (Multiple Precision Floating point Reliable) Libraries
-
-PKG_LIBS = -lmpfr -lgmp
 CXX_STD = CXX11
