+++ src/doubleWishart.cpp
@@ -1,28 +1,41 @@
 #include <Rcpp.h>
 #include <Eigen/Dense>
 #include <Eigen/LU>
-#include "utility.h"
 
+// [[Rcpp::depends(BH)]]
+
+#include <boost/math/special_functions/beta.hpp>
+#include <boost/math/special_functions/gamma.hpp>
+#include <boost/multiprecision/cpp_dec_float.hpp>
+
+using namespace boost::multiprecision;
 using namespace Rcpp;
 using namespace Eigen;
+using boost::math::constants::pi;
 
-// [[Rcpp::export]]
-double doubleWishart_C(double x, int s, double m, double n) {
+typedef number<cpp_dec_float<100>> mp_float;
+
+// Declare matrix and vector types with multi-precision scalar type
+// typedef Eigen::Matrix<mp_float,Dynamic,Dynamic>  MatrixXmp;
+// typedef Eigen::Matrix<mp_float,Dynamic,1>        VectorXmp;
+
+
+template <class T>
+T doubleWishart(T xx, int s, T mm, T nn) {
     // Initialize vector
-    VectorXd b(s);
+    Eigen::Matrix<T,Dynamic,1> b(s);
     int d = s + (s % 2);
 
     // Initialize matrix with zeroes on diagonal
-    MatrixXd A(d, d);
+    Eigen::Matrix<T,Dynamic,Dynamic> A(d, d);
     for (int i = 0; i < d; i++) {
         A(i, i) = 0;
     }
 
-
     if (s != d) {
         // Fill in extra column
         for (int i = 0; i < s; i++) {
-            A(i, s) = incompleteBeta_C(x, m + i + 1, n + 1);
+            A(i, s) = boost::math::beta(mm + i + 1, nn + 1, xx);
             A(s, i) = - A(i, s);
         }
     }
@@ -30,11 +43,11 @@ double doubleWishart_C(double x, int s, double m, double n) {
 
     if (s != 1) {
         for (int i = 0; i < s; i++) {
-            b(i) = 0.5 * pow(incompleteBeta_C(x, m + i + 1, n + 1), 2);
+            b(i) = 0.5 * pow(boost::math::beta(mm + i + 1, nn + 1, xx), 2);
 
             for (int j = i; j < (s-1); j++) {
-                b(j+1) = ((m+j+1)*b(j) - incompleteBeta_C(x, 2*m+i+j+2, 2*n+2))/(m+j+n+2);
-                A(i,j+1) = incompleteBeta_C(x, m+i+1, n+1) * incompleteBeta_C(x, m+j+2, n+1) -
+                b(j+1) = ((mm+j+1)*b(j) - boost::math::beta(2*mm+i+j+2, 2*nn+2, xx))/(mm+j+nn+2);
+                A(i,j+1) = boost::math::beta(mm+i+1, nn+1, xx) * boost::math::beta(mm+j+2, nn+1, xx) -
                     2*b(j+1);
                 A(j+1, i) = - A(i, j+1);
             }
@@ -43,21 +56,37 @@ double doubleWishart_C(double x, int s, double m, double n) {
 
 
     // Compute scaling constant
-    double C1 = 0;
+    T C1 = 0;
     for (int i = 1; i <= s; i++) {
-        C1 += lgamma(0.5*(i+2*m+2*n+s+2)) - lgamma(0.5*i) -lgamma(0.5*(i+2*m+1)) - lgamma(0.5*(i+2*n+1));
+        C1 += boost::math::lgamma(0.5*(i+2*mm+2*nn+s+2)) - boost::math::lgamma(0.5*i) -
+            boost::math::lgamma(0.5*(i+2*mm+1)) - boost::math::lgamma(0.5*(i+2*nn+1));
     }
-    double C = pow(M_PI, 0.5 * s) * exp(C1);
+    T C = pow(pi<T>(), 0.5 * s) * exp(C1);
 
-    double det;
+    T det;
     if(d > 4) {
         det = A.fullPivLu().determinant();
     } else {
         det = A.determinant();
     }
 
-    double result = C * sqrt(det);
+    T result = C * sqrt(det);
     return result;
+}
 
+// [[Rcpp::export]]
+double doubleWishart_raw(double x, int s, double m, double n, bool mp) {
+    double result;
+    if (mp) {
+        mp_float xx(x);
+        mp_float mm(m);
+        mp_float nn(n);
+
+        result = doubleWishart(xx, s, mm, nn).convert_to<double>();
+    } else {
+        result = doubleWishart(x, s, m, n);
+    }
+
+    return result;
 }
