--- src/utility.h
@@ -1,8 +1,0 @@
-#ifndef UTILITY_H
-#define UTILITY_H
-
-double incompleteBeta_C(double x, double alpha, double beta);
-
-double mgamma_C(double, int, bool);
-
-#endif
